% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltjmm-package.R
\docType{package}
\name{ltjmm-package}
\alias{ltjmm-package}
\title{ltjmm: Latent Time Joint Mixed Effect Models}
\description{
Diseases that progress over long periods of time are often studied by observing cohorts at different stages of disease for shorter periods of time. We apply MCMC sampling to estimate Latent Time Joint Mixed Effect Models (LTJMM) from short-term observations with unknown relative observation times.
}
\author{
\strong{Maintainer}: Michael C Donohue \email{mdonohue@usc.edu} (\href{https://orcid.org/0000-0001-6026-2238}{ORCID})

Authors:
\itemize{
  \item Dan Li \email{lidan860922@gmail.com}
}

}
\keyword{internal}
