% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltjmm.R
\name{ltjmm}
\alias{ltjmm}
\title{LTJMM data setup for fitting with rstan}
\usage{
ltjmm(formula, data, subset, na.action)
}
\arguments{
\item{formula}{a \code{\link[Formula]{Formula}} with \code{rhs} denoted the outcome in the 
stacked dataset, and \code{rhs} with 4 parts, e.g.: \code{Y ~ variable for observation time | fixed
effects | subject id | outcome id}}

\item{data}{data.frame containing the variables in the model.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain NAs.}
}
\description{
LTJMM data setup for fitting with rstan
}
\seealso{
\code{\link[rstan]{stan}}
}
