% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltjmm_stan.R
\name{ltjmm_stan}
\alias{ltjmm_stan}
\title{Latent time joint mixed effect model with univariate normal distributions on random effects fit with stan}
\usage{
ltjmm_stan(
  formula,
  lt = TRUE,
  random_effects = "univariate",
  data,
  subset,
  na.action,
  ...
)
}
\arguments{
\item{formula}{a \code{\link[Formula]{Formula}} with \code{rhs} denoted the outcome in the 
stacked dataset, and \code{rhs} with 4 parts, e.g.: \code{Y ~ variable for observation time | fixed
effects | subject id | outcome id}}

\item{lt}{logical, indicating whether or not latent time effect should be included.}

\item{random_effects}{character specifying distribution for random intercepts and slopes. Option 'univariate' 
specifies that random intercepts and slopes for each outcome follow univariate independent normal distributions.
Option 'multivariate' specifies that random intercepts and slopes follow a single multivariate normal distribution.
Option 'multivariate_no_constraint' with \code{lt=TRUE} relaxes the sum-to-zero constraint described in
Li et al. 2017. Option 'none' drops random intercepts and slopes from the model.}

\item{data}{data.frame containing the variables in the model.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain NAs.}

\item{...}{Arguments passed to `cmdstanr::sample` (e.g. iter_warmup, iter_sampling, chains, parallel_chains).}
}
\value{
A \code{\link[cmdstanr]{CmdStanMCMC}} object.
}
\description{
Latent time joint mixed effect model with univariate normal distributions on random effects fit with stan
}
\seealso{
\code{\link[ltjmm]{ltjmm}}
}
