% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltjmm.R
\name{predict.ltjmm}
\alias{predict.ltjmm}
\title{Obtain predictions and confidence intervals for LTJMM models}
\usage{
\method{predict}{ltjmm}(
  object,
  stanfit.object,
  newdata,
  link,
  level = 0.95,
  grouping = c("population", "subject")[1],
  ...
)
}
\arguments{
\item{object}{a \code{\link[ltjmm]{ltjmm}} object}

\item{stanfit.object}{a corresponding \code{\link[cmdstanr]{CmdStanMCMC}} object}

\item{newdata}{new data on which to generate predictions}

\item{link}{link function (defaults to identity)}

\item{level}{for credible intervals}

\item{grouping}{character indicating whether 'population' or 'subject' level precictions
are desired.}

\item{...}{additional optional arguments.}
}
\description{
Obtain predictions and confidence intervals for LTJMM models
}
\seealso{
\code{\link[ltjmm]{ltjmm}}
}
