% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltjmm.R
\name{simulate.ltjmm}
\alias{simulate.ltjmm}
\title{Simulate data from LTJMM with multivariate normal distribution for random effects}
\usage{
\method{simulate}{ltjmm}(
  object,
  nsim = object$data$N_obs,
  seed = NULL,
  beta = array(1, c(object$data$N_out, object$data$N_X)),
  gamma = rep(1, object$data$N_out),
  sigma_delta = 1,
  sigma_y = rep(1, object$data$N_out),
  sigma_diag = diag(rep(1, 2 * object$data$N_out - 1)),
  Lcorr = diag(rep(1, 2 * object$data$N_out - 1)),
  delta = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{\link[ltjmm]{ltjmm}} object}

\item{nsim}{number of response vectors to simulate. Defaults to 1.}

\item{seed}{random seed}

\item{beta}{fixed effects for covariates}

\item{gamma}{latent time slope}

\item{sigma_delta}{standard deviation for latent time}

\item{sigma_y}{standard deviation for residual variance}

\item{sigma_diag}{(Cholesky factorization diag_matrix(sigma_diag) * Lcorr ) for random intercepts and slopes}

\item{Lcorr}{(Cholesky factorization diag_matrix(sigma_diag) * Lcorr) for random intercepts and slopes}

\item{delta}{vector of latent times. If NULL, latent times delta are simulated by normal(0, sigma_delta).}

\item{...}{additional optional arguments.}
}
\description{
Simulate data from LTJMM with multivariate normal distribution for random effects
}
\seealso{
\code{\link[ltjmm]{ltjmm}}
}
